file structure refactor; comments removed; plan to reintroduce
