+++ src/main.rs
@@ -1,9 +1,12 @@
+// src/main.rs
+
 mod log_data;
 mod constants;
-mod plotting_utils;
+mod plot_framework;  // Shared plotting logic
+mod plot_functions; // Individual plot functions
 mod step_response;
 mod fft_utils;
-mod log_parser; // New import
+mod log_parser;
 
 use std::error::Error;
 use std::env;
@@ -12,15 +15,14 @@ use std::path::Path;
 use ndarray::{Array1, Array2};
 
 use constants::*;
-// Import the specific plot functions
-use plotting_utils::{
-    plot_pidsum_error_setpoint,
-    plot_setpoint_vs_gyro,
-    plot_gyro_vs_unfilt,
-    plot_step_response,
-    plot_gyro_spectrums, // New import for the gyro spectrum plot
-};
-use log_parser::parse_log_file; // Import the new parsing function
+// Import the specific plot functions from their new locations
+use plot_functions::plot_pidsum_error_setpoint::plot_pidsum_error_setpoint;
+use plot_functions::plot_setpoint_vs_gyro::plot_setpoint_vs_gyro;
+use plot_functions::plot_gyro_vs_unfilt::plot_gyro_vs_unfilt;
+use plot_functions::plot_step_response::plot_step_response;
+use plot_functions::plot_gyro_spectrums::plot_gyro_spectrums;
+
+use log_parser::parse_log_file;
 
 fn main() -> Result<(), Box<dyn Error>> {
     // --- Argument Parsing ---
@@ -40,8 +42,8 @@ fn main() -> Result<(), Box<dyn Error>> {
         sample_rate,
         setpoint_header_found,
         gyro_header_found,
-        _gyro_unfilt_header_found, // Not directly used here, but returned by parser
-        _debug_header_found,       // Not directly used here, but returned by parser
+        _gyro_unfilt_header_found,
+        _debug_header_found,
     ) = parse_log_file(&input_path)?;
 
     if all_log_data.is_empty() {
@@ -50,10 +52,6 @@ fn main() -> Result<(), Box<dyn Error>> {
     }
 
     // --- Calculate Step Response Data ---
-    // Prepare step response input data filtered by time and movement threshold *once*.
-    // This filtering logic is now moved inside calculate_step_response.
-    // This needs first_time and last_time which are implicitly available from the first/last row,
-    // and requires setpoint and gyro being available in the log data.
     let mut contiguous_sr_input_data: [(Vec<f64>, Vec<f32>, Vec<f32>); 3] = [
         (Vec::new(), Vec::new(), Vec::new()),
         (Vec::new(), Vec::new(), Vec::new()),
@@ -63,18 +61,16 @@ fn main() -> Result<(), Box<dyn Error>> {
     let first_time = all_log_data.first().and_then(|row| row.time_sec);
     let last_time = all_log_data.last().and_then(|row| row.time_sec);
 
-    // Only attempt to filter input data if time range and sample rate are available AND required headers were found
     let mut required_headers_for_sr_input = true;
-    for axis in 0..3 { // Only check setpoint[0-2] for step response input
+    for axis in 0..3 {
         if !setpoint_header_found[axis] || !gyro_header_found[axis] {
             required_headers_for_sr_input = false;
             break;
         }
     }
 
-    if let (Some(first_time_val), Some(last_time_val), Some(_sr)) = (first_time, last_time, sample_rate) { // _sr is unused here
+    if let (Some(first_time_val), Some(last_time_val), Some(_sr)) = (first_time, last_time, sample_rate) {
          if required_headers_for_sr_input {
-              // Collect all data within the time exclusion range
              for row in &all_log_data {
                  if let (Some(time), Some(setpoint_roll), Some(gyro_roll),
                          Some(setpoint_pitch), Some(gyro_pitch),
@@ -102,15 +98,11 @@ fn main() -> Result<(), Box<dyn Error>> {
          println!("\nINFO: Skipping Step Response input data filtering: {}.", reason);
     }
 
-
     println!("\n--- Calculating Step Response ---");
-    // Store the raw QC'd stacked responses and setpoints for later averaging *within the plot function*.
     let mut step_response_calculation_results: [Option<(Array1<f64>, Array2<f32>, Array1<f32>)>; 3] = [None, None, None];
 
      if let Some(sr) = sample_rate {
         for axis_index in 0..3 {
-            // Check if there's *any* movement-filtered data for this axis AND required headers were found
-            // The movement filtering is now inside the calculation function
             let required_headers_found = setpoint_header_found[axis_index] && gyro_header_found[axis_index];
             if required_headers_found && !contiguous_sr_input_data[axis_index].0.is_empty() {
                 println!("  Attempting step response calculation for Axis {}...", axis_index);
@@ -122,9 +114,9 @@ fn main() -> Result<(), Box<dyn Error>> {
                 if time_arr.len() >= min_required_samples {
                     match step_response::calculate_step_response(&time_arr, &setpoints_arr, &gyros_filtered_arr, sr) {
                         Ok(result) => {
-                             let num_qc_windows = result.1.shape()[0]; // Check number of QC windows
+                             let num_qc_windows = result.1.shape()[0];
                              if num_qc_windows > 0 {
-                                 step_response_calculation_results[axis_index] = Some(result); // Store the calculated data.
+                                 step_response_calculation_results[axis_index] = Some(result);
                                  println!("    ... Calculation successful for Axis {}. {} windows passed QC.", axis_index, num_qc_windows);
                              } else {
                                 println!("    ... Calculation returned no valid windows for Axis {}. Skipping.", axis_index);
@@ -140,7 +132,7 @@ fn main() -> Result<(), Box<dyn Error>> {
             } else {
                  let reason = if !required_headers_found {
                      "Setpoint or Gyro headers missing"
-                 } else { // This case should be caught by the time range check above, but kept for clarity
+                 } else {
                      "No movement-filtered input data available"
                  };
                  println!("  Skipping Step Response calculation for Axis {}: {}", axis_index, reason);
@@ -150,17 +142,12 @@ fn main() -> Result<(), Box<dyn Error>> {
          println!("  Skipping Step Response Calculation: Sample rate could not be determined.");
     }
 
-
     // --- Generate Plots ---
-    // Pass the full log_data and root_name to each plotting function
     plot_pidsum_error_setpoint(&all_log_data, &root_name)?;
     plot_setpoint_vs_gyro(&all_log_data, &root_name)?;
     plot_gyro_vs_unfilt(&all_log_data, &root_name)?;
-    // Pass step response results and sample rate to the step response plot function
     plot_step_response(&step_response_calculation_results, &root_name, sample_rate)?;
-    // New: Plot Gyro Spectrums
     plot_gyro_spectrums(&all_log_data, &root_name, sample_rate)?;
-
 
     println!("\nProcessing complete.");
     Ok(())
