+++ src/plot_functions/mod.rs
@@ -1,0 +1,9 @@
+// src/plot_functions/mod.rs
+
+pub mod plot_pidsum_error_setpoint;
+pub mod plot_setpoint_vs_gyro;
+pub mod plot_gyro_vs_unfilt;
+pub mod plot_step_response;
+pub mod plot_gyro_spectrums;
+
+// src/plot_functions/mod.rs
